+++ cuneiform_src/Kern/rstr/src/tools.c
@@ -392,7 +392,7 @@ version *pv;
 extern uchar *CellsPage_rstr, *CellsPageEnd_rstr;
 void setup_string()
 {
-int32_t    number_of_cells=((int32_t)CellsPageEnd_rstr-(int32_t)CellsPage_rstr)/sizeof (cell);
+int32_t    number_of_cells=((intptr_t)CellsPageEnd_rstr-(intptr_t)CellsPage_rstr)/sizeof (cell);
  empty_cell=(cell *) (CellsPage_rstr);
  cell_boundary = empty_cell + number_of_cells;//NUMBER_OF_CELLS;
  free_cell_chain=NULL;
